requirements.txt
torch>=2.0.0
torchvision
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
xformers
bitsandbytes
torchmetrics
pytorch-fid
swanlab
numpy
Pillow
tqdm